换数学库，接入GLFW

+++ tgaimage.h
@@ -43,7 +43,7 @@ struct TGAColor
 	{
 		TGAColor res = *this;
 		double clamped = std::max(0., std::min(intensity, 1.));
-		for (int i = 0; i < 4; i++) res.bgra[i] = bgra[i] * clamped;
+		for (int i = 0; i < 4; i++) res.bgra[i] = bgra[i] * (uint8_t)clamped;
 		return res;
 	}
 };
